Move threshold to shared file
