+++ src/applications/pensions/config/chapters/05-financial-information/netWorthEstimationFormNeeded.js
@@ -1,6 +1,6 @@
 import { titleUI } from 'platform/forms-system/src/js/web-component-patterns';
 import { NetWorthEstimationFormNeededDescription } from '../../../components/Descriptions';
-import { hideIfUnderThreshold } from './helpers';
+import { hideIfUnderThreshold } from '../../../helpers';
 
 export default {
   title: 'Additional form needed',
