+++ src/applications/pensions/config/chapters/05-financial-information/totalNetWorth.js
@@ -4,11 +4,9 @@ import {
 } from 'platform/forms-system/src/js/web-component-patterns';
 import fullSchemaPensions from 'vets-json-schema/dist/21P-527EZ-schema.json';
 import { AssetsInformation } from '../../../components/FormAlerts';
-import { showPdfFormAlignment } from '../../../helpers';
+import { threshold } from '../../../helpers';
 
 const { totalNetWorth } = fullSchemaPensions.properties;
-
-const threshold = showPdfFormAlignment() ? 75000 : 25000;
 
 /** @type {PageSchema} */
 export default {
